MultiThreaded Server + client confirmations and disconnection
Multithreaded Server. 
Now Server just accept the connections and delegate all the communication between the single client to ServerConnection.
Also the Server maintains the methods to modify the status of the login phase.
Client now for number of players and game mode can confirm the already set values or disconnect gracefully