Lobby returns MessageParsers + CLI support now more than only login
Now when the lobby is full, the ClientHandlers receive a notify and they try to get theirs MessageParsers to communicate directly to the controller.
MessageParser given the message translate it to method use of the GameHandler created by the Lobby.
Now the CLI can go beyond the login and initiate all the nominal procedures and message exchanges