First specialCard usage method + SpecialCard phases + messages
ExpertGameOrchestrator has now the first mehtod that updates the GameOrchestrator state to wait for the correct input messages needed for using the specialCards. 
Also there is now an answer message method in MessageGenerator that requires the next parameter to use the special card correctly such as choose a color or an island.