SpecialCard: chooseColor implemented
The chooseColor method is now implemented. 
It functionality is needed in order to take the needed color for the special card usage. 
An interaction for the specialCard usage can need more methods, such as useSpecialCard to activate the card, then chooseColor to choose the color from the card and then chooseIsland to place that student into the Island.