WIP - CurrentPhase + CurrentPlayer updates correction View State
Now, the ViewState has a currentPhase that active the View if we are the active player and disable it if we are not.
Now the View State has the current player to check if the player is active in a specified phase.
Model - GameBoard now have listenable extension needed to then do all the updates.