CLI + ViewStates changes to synchronize and manage planning phase
This commit main goal is to set correctly the assistant card use phase. Now the two player can play a card.
The CLI now awaits the ViewState for be awaken with wait/notify to use less resources and to do not block with synchronized method the ViewMessageParser.
The clients into GameBoard were set as the same List of Lobby. This was degenerative since the clients then where removed from the Lobby queue making impossible the correct updates. Now this problem is solved.