Gui - new version of LobbyMenu players adding
I changed the "update flow" in the Lobby scene:  now as soon as  the ViewMessage receives a "Player Update" messages, it calls the MainGUI which in turn calls the LobbyMenu controller and updates the online players list.  Finally the MainGUI reloads the scene with setNextStage.
There's no need for the previous scenes (setup and acceptCondition) to set the next scene: it is handled by ViewMessageParser and MainGUI. :)