Structural changes CLI, now keyboard has its thread, print by viewState
Now the print is done by the viewState when a new message change it state.
Using a subPhase enumeration, it is possible to know the correct instructions to print.
The input now is all manage by the KeyboardInputReader class after the login.
KeyboardInputReader just waits for input and consume it, controlling if the view is active and in that case what input is expected in that particular phase, then or it sends the message or memorize the intermediate information to then complete them and send the message to the Server