Some test issues + cloud usage issue solved
There were some issue with tests due to a notify done after a counter increased that made the notify take an out of range value. 
Also the PlayerBoards begins with 1 coin and not 2, fixed.
Cloud usage improved with a flag that tells if the cloud has been used, this allow to avoid situations where the cloud is wrongly taken as already used but it is just not filled because of the empty bag