Issue with disconnection timer - lazy ping made disconnect
The ping was sent only if no message received. By doing so, the time to send the next ping was too much and made disconnect server side every client. 
Also, the timer to disconnect a particular client, server side should be stopped to avoid disconnecting that client 2 times even if it come back with no connection issues