CLI prints improvement + last used card now updated + port only num
Now the last card is correctly updated to the viewState
Graphical improvement in the print of the playerBoards 
Now the playerBoards are printed for the moveMotherNature and for the cloud choice to know the current situation and the professor ownership.
Islands are now printed first than use the assistant card and choice the clouds.
The port now, if the person insert a string and not a number does not make crash all